refactor: move KIS exceptions to temp package and update imports
