+++ src/main/java/co/fineants/api/domain/kis/repository/ClosingPriceRepository.java
@@ -10,9 +10,9 @@ import co.fineants.api.domain.common.money.Money;
 import co.fineants.api.domain.kis.client.KisClient;
 import co.fineants.api.domain.kis.domain.dto.response.KisClosingPrice;
 import co.fineants.api.global.common.delay.DelayManager;
-import co.fineants.api.global.errors.exception.kis.CredentialsTypeKisException;
-import co.fineants.api.global.errors.exception.kis.ExpiredAccessTokenKisException;
-import co.fineants.api.global.errors.exception.kis.RequestLimitExceededKisException;
+import co.fineants.api.global.errors.exception.temp.CredentialsTypeKisException;
+import co.fineants.api.global.errors.exception.temp.ExpiredAccessTokenKisException;
+import co.fineants.api.global.errors.exception.temp.RequestLimitExceededKisException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import reactor.core.Exceptions;
