+++ src/main/java/co/fineants/api/domain/kis/repository/CurrentPriceRedisRepository.java
@@ -13,9 +13,9 @@ import co.fineants.api.domain.kis.client.KisClient;
 import co.fineants.api.domain.kis.client.KisCurrentPrice;
 import co.fineants.api.domain.stock.domain.entity.Stock;
 import co.fineants.api.global.common.delay.DelayManager;
-import co.fineants.api.global.errors.exception.kis.CredentialsTypeKisException;
-import co.fineants.api.global.errors.exception.kis.ExpiredAccessTokenKisException;
-import co.fineants.api.global.errors.exception.kis.RequestLimitExceededKisException;
+import co.fineants.api.global.errors.exception.temp.CredentialsTypeKisException;
+import co.fineants.api.global.errors.exception.temp.ExpiredAccessTokenKisException;
+import co.fineants.api.global.errors.exception.temp.RequestLimitExceededKisException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import reactor.core.Exceptions;
