+++ src/main/java/co/fineants/api/global/errors/exception/temp/KisApiRequestException.java
@@ -3,8 +3,6 @@ package co.fineants.api.global.errors.exception.temp;
 import org.springframework.http.HttpStatus;
 
 import co.fineants.api.global.errors.errorcode.ErrorCode;
-import co.fineants.api.global.errors.exception.kis.ExpiredAccessTokenKisException;
-import co.fineants.api.global.errors.exception.kis.RequestLimitExceededKisException;
 
 public class KisApiRequestException extends BusinessException {
 	private final String returnCode;
